Added option --no-fixrelative
When exporting to HTML, the `src` attributes were by default modified to
change all relative routes to absolute ones.

Unfortunately this also happened inside html code blocks, so I included
a switch to disable this behaviour.

Relative routes can cause problems however in the "main" html, but not
when resources are inlined.